+++ app/charts/index.ts
@@ -293,7 +293,7 @@ export const getInitialConfig = ({
         },
         baseLayer: {
           showRelief: true,
-          showLakes: true,
+          showWater: true,
         },
       };