+++ app/configurator/components/chart-configurator.tsx
@@ -431,8 +431,8 @@ const ChartFields = ({
             icon="baseLayer"
             label={<Trans id="chart.map.layers.base">Base Layer</Trans>}
             active={
-              chartConfig.baseLayer.showLakes ||
-              chartConfig.baseLayer.showRelief
+              chartConfig.baseLayer.showRelief ||
+              chartConfig.baseLayer.showWater
             }
           />
         ) : (